requirements.txt
ecdsa
cryptography>3.1
bluepy
RPi.GPIO